requirements.txt
Django==3.2.7
djangorestframework==3.12.4
psycopg2-binary==2.9.7
django-filter==23.2
Pillow==10.0